requirements.txt
Flask==2.1.1
torch==1.10.0
torchvision==0.11.1
requests==2.22.0
Pillow==7.0.0
gunicorn==20.1.0
prometheus-flask-exporter==0.18.7
pytest
grpcio-tools
grpcio
prometheus-client
